refactor: move orchestration services from core to app layer
