--- 1 - Aplicacao/Sistema.APP/Services/EmailAppService.cs
@@ -1,11 +1,0 @@
-using Sistema.APP.Services.Interfaces;
-
-namespace Sistema.APP.Services;
-
-public class EmailAppService(Sistema.CORE.Services.Interfaces.IEmailDomainService domainService) : IEmailAppService
-{
-    private readonly Sistema.CORE.Services.Interfaces.IEmailDomainService _domainService = domainService;
-
-    public Task EnviarAsync(string destinatario, string assunto, string mensagem, CancellationToken cancellationToken = default) =>
-        _domainService.EnviarAsync(destinatario, assunto, mensagem, cancellationToken);
-}
