+++ 1 - Aplicacao/Sistema.APP/Services/FuncionalidadeAppService.cs
@@ -1,25 +1,41 @@
 using Sistema.APP.Services.Interfaces;
 using Sistema.CORE.Common;
 using Sistema.CORE.Entities;
+using Sistema.CORE.Repositories.Interfaces;
 
 namespace Sistema.APP.Services;
 
-public class FuncionalidadeAppService(Sistema.CORE.Services.Interfaces.IFuncionalidadeDomainService domainService) : IFuncionalidadeAppService
+public class FuncionalidadeAppService(IUnitOfWork uow, ILogAppService log) : IFuncionalidadeAppService
 {
-    private readonly Sistema.CORE.Services.Interfaces.IFuncionalidadeDomainService _domainService = domainService;
+    private readonly IUnitOfWork _uow = uow;
+    private readonly ILogAppService _log = log;
 
-    public Task<PagedResult<Funcionalidade>> BuscarPaginadasAsync(int page, int pageSize, CancellationToken cancellationToken = default) =>
-        _domainService.BuscarPaginadasAsync(page, pageSize, cancellationToken);
+    public Task<PagedResult<Funcionalidade>> BuscarPaginadasAsync(int page, int pageSize, CancellationToken cancellationToken = default)
+        => _uow.Funcionalidades.BuscarPaginadasAsync(page, pageSize, cancellationToken);
 
-    public Task<Funcionalidade?> BuscarPorIdAsync(int id, CancellationToken cancellationToken = default) =>
-        _domainService.BuscarPorIdAsync(id, cancellationToken);
+    public Task<Funcionalidade?> BuscarPorIdAsync(int id, CancellationToken cancellationToken = default) => _uow.Funcionalidades.BuscarPorIdAsync(id, cancellationToken);
 
-    public Task<OperationResult<Funcionalidade>> AdicionarAsync(Funcionalidade func, CancellationToken cancellationToken = default) =>
-        _domainService.AdicionarAsync(func, cancellationToken);
+    public async Task<OperationResult<Funcionalidade>> AdicionarAsync(Funcionalidade func, CancellationToken cancellationToken = default)
+    {
+        await _uow.Funcionalidades.AdicionarAsync(func, cancellationToken);
+        await _log.RegistrarAsync(nameof(Funcionalidade), "Add", true, "Funcionalidade criada", LogTipo.Sucesso, func.UsuarioInclusao, null, cancellationToken);
+        await _uow.ConfirmarAsync(cancellationToken);
+        return new OperationResult<Funcionalidade>(true, "Criado", func);
+    }
 
-    public Task<OperationResult> AtualizarAsync(Funcionalidade func, CancellationToken cancellationToken = default) =>
-        _domainService.AtualizarAsync(func, cancellationToken);
+    public async Task<OperationResult> AtualizarAsync(Funcionalidade func, CancellationToken cancellationToken = default)
+    {
+        await _uow.Funcionalidades.AtualizarAsync(func);
+        await _log.RegistrarAsync(nameof(Funcionalidade), "Update", true, "Funcionalidade atualizada", LogTipo.Sucesso, func.UsuarioAlteracao ?? "system", null, cancellationToken);
+        await _uow.ConfirmarAsync(cancellationToken);
+        return new OperationResult(true, "Atualizado");
+    }
 
-    public Task<OperationResult> RemoverAsync(int id, CancellationToken cancellationToken = default) =>
-        _domainService.RemoverAsync(id, cancellationToken);
+    public async Task<OperationResult> RemoverAsync(int id, CancellationToken cancellationToken = default)
+    {
+        await _uow.Funcionalidades.RemoverAsync(id, cancellationToken);
+        await _log.RegistrarAsync(nameof(Funcionalidade), "Delete", true, "Funcionalidade removida", LogTipo.Sucesso, "system", null, cancellationToken);
+        await _uow.ConfirmarAsync(cancellationToken);
+        return new OperationResult(true, "Removido");
+    }
 }
