--- 2 - Dominio/Sistema.CORE/Services/Interfaces/IEmailDomainService.cs
@@ -1,9 +1,0 @@
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Sistema.CORE.Services.Interfaces;
-
-public interface IEmailDomainService
-{
-    Task EnviarAsync(string destinatario, string assunto, string mensagem, CancellationToken cancellationToken = default);
-}
