+++ 3 - Infraestrutura/Sistema.INFRA/Services/EmailService.cs
@@ -4,12 +4,12 @@ using Microsoft.Extensions.Options;
 using Microsoft.Graph;
 using Microsoft.Graph.Models;
 using Microsoft.Graph.Users.Item.SendMail;
-using Sistema.CORE.Services.Interfaces;
+using Sistema.APP.Services.Interfaces;
 using System.Threading;
 
 namespace Sistema.INFRA.Services;
 
-public class EmailService(IOptions<EmailOptions> options, ILogger<EmailService> logger) : IEmailDomainService
+public class EmailService(IOptions<EmailOptions> options, ILogger<EmailService> logger) : IEmailAppService
 {
     private readonly EmailOptions _options = options.Value;
     private readonly ILogger<EmailService> _logger = logger;
