+++ 3 - Infraestrutura/Sistema.INFRA/Sistema.INFRA.csproj
@@ -1,6 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <ItemGroup>
+    <ProjectReference Include="..\..\1 - Aplicacao\Sistema.APP\Sistema.APP.csproj" />
     <ProjectReference Include="..\..\2 - Dominio\Sistema.CORE\Sistema.CORE.csproj" />
   </ItemGroup>
